moved test to existing autograd test file
